arrange libs
