+++ .env.example
@@ -11,4 +11,4 @@ NEXT_PUBLIC_GISCUS_CATEGORY_ID=XXX
 NEXT_PUBLIC_GA_ID=XXX
 
 # Backend Database URL
-ATMAN_DB_URL=mongodb://sampleuser:pass@localhost:27017/sampledb
+ATMAN_DB_URL=mongodb://sampleuser:pass@localhost/sampledb
