+++ docs/manual/03-nest-graphql-apollo-prisma/02-generate-db-types.md
@@ -3,7 +3,7 @@
 ## generate types
 
 ```bash
-npx nx generate @nx/js:library api-generated-db-types --directory=libs/api/generated-db-types --importPath=@libs/api/generated-db-types --tags=scope:api --bundler=swc
+npx nx generate @nx/js:library api-generated-db-types --directory=libs/api/prisma/generated-db-types --importPath=@libs/api/prisma/generated-db-types --tags=scope:api --bundler=swc
 
 ✔ Which unit test runner would you like to use? · none
 ```
@@ -18,7 +18,7 @@ npm i -D prisma-nestjs-graphql
 
 ### setting generator nestgraphql to schema.prisma
 
- `libs/api/data-access-db/src/lib/schema.prisma`
+ `libs/api/prisma/data-access-db/src/lib/schema.prisma`
 
 # For Frontend
