--- libs/api/feature-sample/README.md
@@ -1,11 +1,0 @@
-# api-feature-sample
-
-This library was generated with [Nx](https://nx.dev).
-
-## Building
-
-Run `nx build api-feature-sample` to build the library.
-
-## Running unit tests
-
-Run `nx test api-feature-sample` to execute the unit tests via [Jest](https://jestjs.io).
