--- libs/api/feature-sample/package.json
@@ -1,10 +1,0 @@
-{
-  "name": "@libs/api/feature-sample",
-  "version": "0.0.1",
-  "dependencies": {
-    "@swc/helpers": "~0.5.0"
-  },
-  "type": "commonjs",
-  "main": "./src/index.js",
-  "typings": "./src/index.d.ts"
-}
