--- libs/api/feature-sample/project.json
@@ -1,43 +1,0 @@
-{
-  "name": "api-feature-sample",
-  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/api/feature-sample/src",
-  "projectType": "library",
-  "targets": {
-    "build": {
-      "executor": "@nx/js:swc",
-      "outputs": ["{options.outputPath}"],
-      "options": {
-        "outputPath": "dist/libs/api/feature-sample",
-        "main": "libs/api/feature-sample/src/index.ts",
-        "tsConfig": "libs/api/feature-sample/tsconfig.lib.json",
-        "assets": ["libs/api/feature-sample/*.md"]
-      }
-    },
-    "lint": {
-      "executor": "@nx/linter:eslint",
-      "outputs": ["{options.outputFile}"],
-      "options": {
-        "lintFilePatterns": [
-          "libs/api/feature-sample/**/*.ts",
-          "libs/api/feature-sample/package.json"
-        ]
-      }
-    },
-    "test": {
-      "executor": "@nx/jest:jest",
-      "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
-      "options": {
-        "jestConfig": "libs/api/feature-sample/jest.config.ts",
-        "passWithNoTests": true
-      },
-      "configurations": {
-        "ci": {
-          "ci": true,
-          "codeCoverage": true
-        }
-      }
-    }
-  },
-  "tags": ["scope:api"]
-}
