--- libs/api/generated-db-types/project.json
@@ -1,29 +1,0 @@
-{
-  "name": "api-generated-db-types",
-  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/api/generated-db-types/src",
-  "projectType": "library",
-  "targets": {
-    "build": {
-      "executor": "@nx/js:swc",
-      "outputs": ["{options.outputPath}"],
-      "options": {
-        "outputPath": "dist/libs/api/generated-db-types",
-        "main": "libs/api/generated-db-types/src/index.ts",
-        "tsConfig": "libs/api/generated-db-types/tsconfig.lib.json",
-        "assets": ["libs/api/generated-db-types/*.md"]
-      }
-    },
-    "lint": {
-      "executor": "@nx/linter:eslint",
-      "outputs": ["{options.outputFile}"],
-      "options": {
-        "lintFilePatterns": [
-          "libs/api/generated-db-types/**/*.ts",
-          "libs/api/generated-db-types/package.json"
-        ]
-      }
-    }
-  },
-  "tags": ["scope:api"]
-}
