+++ libs/api/prisma/data-access-db/README.md
@@ -5,10 +5,3 @@ This library was generated with [Nx](https://nx.dev).
 ## Building
 
 Run `nx build api-data-access-db` to build the library.
-
-## Note
-
-This project is for using prisma.
-
-> Prisma is one of the best approach for using postgres.
-> But not usuful for mongodb.
