+++ libs/api/prisma/data-access-db/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@libs/api/data-access-db",
+  "name": "@libs/api/prisma/data-access-db",
   "version": "0.0.1",
   "dependencies": {
     "@swc/helpers": "~0.5.0"
