+++ libs/api/prisma/data-access-db/project.json
@@ -1,0 +1,29 @@
+{
+  "name": "api-data-access-db",
+  "$schema": "../../../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "libs/api/prisma/data-access-db/src",
+  "projectType": "library",
+  "targets": {
+    "build": {
+      "executor": "@nx/js:swc",
+      "outputs": ["{options.outputPath}"],
+      "options": {
+        "outputPath": "dist/libs/api/prisma/data-access-db",
+        "main": "libs/api/prisma/data-access-db/src/index.ts",
+        "tsConfig": "libs/api/prisma/data-access-db/tsconfig.lib.json",
+        "assets": ["libs/api/prisma/data-access-db/*.md"]
+      }
+    },
+    "lint": {
+      "executor": "@nx/linter:eslint",
+      "outputs": ["{options.outputFile}"],
+      "options": {
+        "lintFilePatterns": [
+          "libs/api/prisma/data-access-db/**/*.ts",
+          "libs/api/prisma/data-access-db/package.json"
+        ]
+      }
+    }
+  },
+  "tags": ["scope:api"]
+}
