+++ libs/api/prisma/data-access-db/src/lib/prisma.module.ts
