+++ libs/api/prisma/generated-db-types/README.md
@@ -5,10 +5,3 @@ This library was generated with [Nx](https://nx.dev).
 ## Building
 
 Run `nx build api-generated-db-types` to build the library.
-
-## Note
-
-This project is for using prisma.
-
-> Prisma is one of the best approach for using postgres.
-> But not usuful for mongodb.
