+++ libs/api/prisma/generated-db-types/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@libs/api/generated-db-types",
+  "name": "@libs/api/prisma/generated-db-types",
   "version": "0.0.1",
   "dependencies": {
     "@swc/helpers": "~0.5.0"
